Updating namespaces to sepearate the _path_ from the ultimate _comand_ run
Mostly logical seperation done so that it's clearer in the code and config files when we're talking about the _path_ to the applications we're running (rtl_sdr or direwofl) or the _command_ we're running to execute that application with the options we want. Also updated the config file to match these names.

This also addresses some issues with Python 2 and Unicode in the direwolf command code from my last commit. Because Unicode. Argh.